requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.2
Django==3.1.6
django-admin-interface==0.16.3
django-appconf==1.0.4
django-celery-email==3.0.0
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-ckeditor-5==0.0.14
django-cleanup==5.2.0
django-colorfield==0.4.2
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-json-widget==1.1.1
django-jsonfield==1.4.1
django-nine==0.2.4
django-rest-framework-tricks==0.2.11
django-rest-passwordreset==1.2.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
dnspython==1.16.0
eventlet==0.31.0
filelock==3.0.12
future==0.18.2
greenlet==1.1.0
idna==2.10
install==1.3.4
kombu==5.1.0
pdfkit==0.6.1
Pillow==8.3.1
platformdirs==2.1.0
prompt-toolkit==3.0.19
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
stripe==2.59.0
urllib3==1.26.6
vine==5.0.0
virtualenv==20.6.0
wcwidth==0.2.5
django-multiselectfield
